Spieler kann sich nur drehen, Filestructure überarbeitet
